Ädding the confirm delete option.

+++ ShopApp/views.py
@@ -3,22 +3,22 @@ from django.http import HttpResponse
 from .models import Productdb
 from .forms import ProductForm
 
-def index(request):
+def create(request):
     if request.method=='POST':
         form=ProductForm(request.POST)
         if form.is_valid():
             try:
                 form.save()
-                return redirect('/show')
+                return redirect('/list')
             except:
                 pass
     else:
         form=ProductForm()
-    return render(request, 'index.html', {'form':form})
+    return render(request, 'create.html', {'form':form})
 
-def show(request):
+def list(request):
     products=Productdb.objects.all()
-    return render(request, 'show.html', {'products':products})
+    return render(request, 'list.html', {'products':products})
 
 def edit(request, id):
     product=Productdb.objects.get(id=id)
@@ -29,13 +29,16 @@ def update(request, id):
     form=ProductForm(request.POST, instance=product)
     if form.is_valid():
         form.save()
-        return redirect("/show")
+        return redirect("/list")
     return render(request, 'edit.html', {'product':product})
 
 def delete(request, id):
     product=Productdb.objects.get(id=id)
-    product.delete()
-    return redirect("/show")
+    if request.method == 'POST': 
+        product.delete() 
+        return redirect("/list")
+    return render(request, 'delete1.html', {'object':product})
+
 
 def delete1(request):
     return render(request, 'delete1.html')
